build without python
While BUILDING claims that building without python is possible,
CMakeLists states python2 as a strict requirement and the build fails
when python2 is not present (even if python3 is). In addition, CMakeLists
in the plugin dir always configures one of the scripters.

Change the CMakeLists files to enable builds without python2 (and
without the scripter).